Moved NRPE related files to the same folder.
